works fine

+++ include/AES.h
@@ -1,9 +1,9 @@
-#include <BigInteger.h>
+#include "../vendor/bigint.h"
 
 class AES
 {
     AES();
 
-    BigInteger encrypt(BigInteger M, BigInteger K);
-    BigInteger decrypt(BigInteger C, BigInteger K);
+    bigint encrypt(bigint M, bigint K);
+    bigint decrypt(bigint C, bigint K);
 };
